tweak: updated the scripts to update the dust levels
