+++ scripts/operations/governance/ypp-0009-updateDust/ypp-0009.ts
@@ -1,0 +1,49 @@
+/**
+ * @dev This script updates the dust limits for one or more base/ilk pairs.
+ *
+ * It takes as inputs the governance and protocol address files.
+ */
+
+import { ethers } from 'hardhat'
+import * as hre from 'hardhat'
+import * as fs from 'fs'
+
+import { getOwnerOrImpersonate, getOriginalChainId, jsonToMap, proposeApproveExecute } from '../../../../shared/helpers'
+import { updateDustProposal } from '../../updateDustProposal'
+import { Cauldron, Timelock } from '../../../../typechain'
+import { newMin } from './ypp-0009.config'
+
+;(async () => {
+
+  const chainId = await getOriginalChainId()
+  if (chainId !== 1 && chainId !== 42) throw 'Only Kovan and Mainnet supported'
+  const path = chainId === 1 ? './addresses/mainnet/' : './addresses/kovan/'
+
+  const developer = new Map([
+    [1, '0xC7aE076086623ecEA2450e364C838916a043F9a8'],
+    [42, '0x5AD7799f02D5a829B2d6FA085e6bd69A872619D5'],
+  ])
+
+  let ownerAcc = await getOwnerOrImpersonate(developer.get(chainId) as string)
+
+  const protocol = jsonToMap(fs.readFileSync(path + 'protocol.json', 'utf8')) as Map<string, string>
+  const governance = jsonToMap(fs.readFileSync(path + 'governance.json', 'utf8')) as Map<string, string>
+
+  // Contract instantiation
+  const cauldron = ((await ethers.getContractAt(
+    'Cauldron',
+    protocol.get('cauldron') as string,
+    ownerAcc
+  )) as unknown) as Cauldron
+
+  const timelock = ((await ethers.getContractAt(
+    'Timelock',
+    governance.get('timelock') as string,
+    ownerAcc
+  )) as unknown) as Timelock
+
+  // Build the proposal
+  const proposal: Array<{ target: string; data: string }> = await updateDustProposal(cauldron, newMin)
+  // Propose, Approve & execute
+  await proposeApproveExecute(timelock, proposal, governance.get('multisig') as string)
+})()
